refactor: Move offset related classes into package
