+++ parallel-consumer-core/src/main/java/io/confluent/csid/utils/JavaUtils.java
@@ -4,7 +4,7 @@ package io.confluent.csid.utils;
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
-import io.confluent.parallelconsumer.InternalRuntimeError;
+import io.confluent.parallelconsumer.internal.InternalRuntimeError;
 
 import java.util.Collection;
 import java.util.List;
