--- parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/BitSetEncodingNotSupportedException.java
@@ -1,12 +1,0 @@
-package io.confluent.parallelconsumer;
-
-/*-
- * Copyright (C) 2020-2021 Confluent, Inc.
- */
-public class BitSetEncodingNotSupportedException extends EncodingNotSupportedException {
-
-    public BitSetEncodingNotSupportedException(String msg) {
-        super(msg);
-    }
-
-}
