--- parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/EncodingNotSupportedException.java
@@ -1,10 +1,0 @@
-package io.confluent.parallelconsumer;
-
-/*-
- * Copyright (C) 2020-2021 Confluent, Inc.
- */
-public class EncodingNotSupportedException extends Exception {
-    public EncodingNotSupportedException(final String message) {
-        super(message);
-    }
-}
