--- parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/OffsetDecodingError.java
@@ -1,10 +1,0 @@
-package io.confluent.parallelconsumer;
-
-/*-
- * Copyright (C) 2020-2021 Confluent, Inc.
- */
-public class OffsetDecodingError extends Exception {
-    public OffsetDecodingError(final String s, final IllegalArgumentException a) {
-        super(s, a);
-    }
-}
