+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/OffsetCommitter.java
@@ -1,4 +1,5 @@
-package io.confluent.parallelconsumer;
+package io.confluent.parallelconsumer.internal;
+
 /*-
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
