+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/offsets/BitSetEncodingNotSupportedException.java
@@ -1,0 +1,12 @@
+package io.confluent.parallelconsumer.offsets;
+
+/*-
+ * Copyright (C) 2020-2021 Confluent, Inc.
+ */
+public class BitSetEncodingNotSupportedException extends EncodingNotSupportedException {
+
+    public BitSetEncodingNotSupportedException(String msg) {
+        super(msg);
+    }
+
+}
