+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/offsets/ByteBufferEncoder.java
@@ -1,4 +1,4 @@
-package io.confluent.parallelconsumer;
+package io.confluent.parallelconsumer.offsets;
 
 /*-
  * Copyright (C) 2020-2021 Confluent, Inc.
@@ -6,8 +6,8 @@ package io.confluent.parallelconsumer;
 
 import java.nio.ByteBuffer;
 
-import static io.confluent.parallelconsumer.OffsetEncoding.ByteArray;
-import static io.confluent.parallelconsumer.OffsetEncoding.ByteArrayCompressed;
+import static io.confluent.parallelconsumer.offsets.OffsetEncoding.ByteArray;
+import static io.confluent.parallelconsumer.offsets.OffsetEncoding.ByteArrayCompressed;
 
 class ByteBufferEncoder extends OffsetEncoder {
