+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/offsets/EncodingNotSupportedException.java
@@ -1,0 +1,10 @@
+package io.confluent.parallelconsumer.offsets;
+
+/*-
+ * Copyright (C) 2020-2021 Confluent, Inc.
+ */
+public class EncodingNotSupportedException extends Exception {
+    public EncodingNotSupportedException(final String message) {
+        super(message);
+    }
+}
