+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/offsets/OffsetEncoder.java
@@ -1,4 +1,4 @@
-package io.confluent.parallelconsumer;
+package io.confluent.parallelconsumer.offsets;
 
 /*-
  * Copyright (C) 2020-2021 Confluent, Inc.
