+++ parallel-consumer-core/src/test/java/io/confluent/csid/utils/KafkaTestUtils.java
@@ -4,7 +4,7 @@ package io.confluent.csid.utils;
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
-import io.confluent.parallelconsumer.OffsetMapCodecManager;
+import io.confluent.parallelconsumer.offsets.OffsetMapCodecManager;
 import io.confluent.parallelconsumer.state.WorkContainer;
 import io.confluent.parallelconsumer.state.WorkManager;
 import lombok.Getter;
