+++ parallel-consumer-core/src/test/java/io/confluent/csid/utils/ProgressBarUtils.java
