+++ parallel-consumer-core/src/test/java/io/confluent/csid/utils/ProgressTracker.java
@@ -4,7 +4,7 @@ package io.confluent.csid.utils;
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
-import io.confluent.parallelconsumer.InternalRuntimeError;
+import io.confluent.parallelconsumer.internal.InternalRuntimeError;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
