+++ parallel-consumer-core/src/test/resources/logback-test.xml
@@ -45,9 +45,9 @@
     <!--    <logger name="io.confluent.parallelconsumer.ParallelConsumerTestBase" level="trace"/>-->
     <!--        <logger name="io.confluent.parallelconsumer.state.WorkManager" level="trace" />-->
     <!--    <logger name="io.confluent.parallelconsumer.ParallelConsumer" level="info" />-->
-    <!--    <logger name="io.confluent.parallelconsumer.OffsetMapCodecManager" level="debug"/>-->
-    <!--    <logger name="io.confluent.parallelconsumer.WorkManagerOffsetMapCodecManagerTest" level="debug"/>-->
-    <!--    <logger name="io.confluent.parallelconsumer.OffsetRunLength" level="trace"/>-->
+    <!--    <logger name="io.confluent.parallelconsumer.offsets.OffsetMapCodecManager" level="debug"/>-->
+    <!--    <logger name="io.confluent.parallelconsumer.offsets.WorkManagerOffsetMapCodecManagerTest" level="debug"/>-->
+    <!--    <logger name="io.confluent.parallelconsumer.offsets.OffsetRunLength" level="trace"/>-->
 
     <!-- containers -->
     <!--    <logger name="io.confluent.parallelconsumer.integrationTests.DbTest" level="debug"/>-->
